pyproject.toml
[project]
name = "evaluating-skills-tutorial"
version = "0.1.0"
description = "Tutorial-sized OpenHands skill evaluation with deterministic tasks and local verifiers."
requires-python = ">=3.12"
dependencies = [
    "openhands-sdk>=1.12.0",
    "openhands-tools>=1.12.0",
    "openhands-workspace>=1.12.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
